refactor: keep curve geometry on the Pen tool
